Updated stage.py to allow stage and unstage diff command
Using the Staged Changes and Unstaged Changes will give the output for git diff [--staged].
These outputs will be saved to the outputs folder within the app. Each file starts with the
name of the current loaded project, a keyword (staged or unstaged to signify which output
it is) and a timestamp with the date and time seperated py a hyphen. Lastly, the creation
of a button that opens the output folder (Open Changes Output). These won't be cleared for
the purpose of tracking changes.