Both commit features added. Addition of run.bat for starting application
Users can now use 'Commit' and 'Quick Commit' to commit work. The commit feature requires users
to type their message in the terminal. The quick commit opens a dialog box where users can type
a short description as a commit note (this uses the -m flag when commmiting). Lastly a batch file
has been added as a makeshift launcher for this project. Eventually an .exe file will be created.